+++ console/Packages/com.cronyx.console/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.2.0] - 2021-03-23
 ### Added
 - Added a public overload for `DeveloperConsole.RegisterCommand` that takes an `IConsoleCommand` instance.
 - Added `VerbCommand`, a new `IConsoleCommand` implementation, with support for verb commands, as in the case of `git push` or `git pull`
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed target exception when calling automatically parsed commands when using non-static methods.
+- Fixed README.md meta file missing error
 
 ## [1.1.1] - 2021-03-23
 ### Changed